Build PLib for netstandard2.1
ONI now uses a cursed version of the Mono runtime that conforms to .NET
Standard 2.1 which does not have an exact Microsoft equivalent. But one
can still target the `netstandard2.1` framework while targetting a .NET
Framework v4 runtime. This is a bit (very) cursed, but somehow works...

I could not get MSBuild to behave the way I wanted it to, so I chose to
override `TargetFrameworks` in the various PLib projects one by one. It
is very likely there is a better way to go about this.

Signed-off-by: Angel Pons <th3fanbus@gmail.com>