PLib: Allow building for .NET Standard 2.1 (#605)
Wrap some stuff in preprocessor as it prevents compilation when using .NET Standard 2.1 (which ONI now uses). The functionality in question seems to be for debugging purposes only so it should not matter much.

ONI now uses a cursed version of the Mono runtime that conforms to .NET Standard 2.1 which does not have an exact Microsoft equivalent. But one can still target the `netstandard2.1` framework while targetting a .NET Framework v4 runtime. This is a bit (very) cursed, but somehow works...

I could not get MSBuild to behave the way I wanted it to, so I chose to override `TargetFrameworks` in the various PLib projects one by one. It is very likely there is a better way to go about this.